feat: added minor improvements
Also added cypress for future testing
