+++ src/components/CostSummary.tsx
@@ -1,4 +1,9 @@
+import { observer } from 'mobx-react-lite';
+import { ParticipantList } from '../data/Actor';
+import { Bill } from '../data/Bill';
 import { $ } from '../helpers/currencyHelper';
+import BillActor from './Actor';
+import { Divider } from './General';
 import Input from './Input';
 
 export interface CostSummaryProps {
@@ -57,5 +62,38 @@ const CostSummary: React.FC<CostSummaryProps> = ({
     </section>
   );
 };
+
+export const CostSummaryView = observer<{
+  billStore: Bill;
+  participantStore: ParticipantList;
+}>(({ billStore, participantStore }) => (
+  <>
+    <CostSummary
+      changeGrand={(a) => billStore.setGrandTotal(a)}
+      changeTaxFees={(a) => billStore.setTaxFees(a)}
+      grandOverride={billStore.grandTotal}
+      subtotal={billStore.subtotal}
+      taxFees={billStore.taxFees}
+    />
+    <Divider />
+    <section>
+      <h3 className="text-3xl font-bold mb-4">Split</h3>
+      {Object.keys(billStore.split).map((key) => {
+        const participant = participantStore.participants.find(
+          (p) => p.id === key,
+        );
+        if (participant) {
+          return (
+            <div key={participant.id} className="flex items-center my-4">
+              <BillActor name={participant.name} size="sm" />
+              <div className="ml-4">{$(billStore.split[key])}</div>
+            </div>
+          );
+        }
+        return <></>;
+      })}
+    </section>
+  </>
+));
 
 export default CostSummary;
